# lxdweb/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateNodeRequest:
    properties:
      address:
        type: string
      api_key:
        type: string
      batch_interval:
        type: integer
      batch_size:
        type: integer
      description:
        type: string
      name:
        type: string
      sync_preset:
        type: string
    required:
    - address
    - name
    type: object
...
  models.UpdateNodeRequest:
    properties:
      address:
        type: string
      api_key:
        type: string
      batch_interval:
        type: integer
      batch_size:
        type: integer
      description:
        type: string
      name:
        type: string
      sync_preset:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: xkatld@example.com
    name: xkatld
    url: https://github.com/xkatld/zjmf-lxd-server
  description: 基于 Session 的认证
  license:
    name: MIT
    url: https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE
  termsOfService: https://github.com/xkatld/zjmf-lxd-server
  title: LXD Web 管理平台
  version: 1.0.3
paths:
...
  /api/nodes:
    get:
      description: 查询所有LXD节点及其系统信息，支持按状态过滤
      parameters:
      - description: 节点状态(active/inactive)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回节点列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties: true
            type: object
      summary: 获取节点列表
      tags:
      - 节点管理
    post:
      consumes:
      - application/json
      description: 添加新的LXD节点到管理系统
      parameters:
      - description: 节点配置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建节点
      tags:
      - 节点管理
  /api/nodes/{id}:
    delete:
      description: 从管理系统中删除指定节点及其所有关联数据
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 删除失败
          schema:
            additionalProperties: true
            type: object
      summary: 删除节点
      tags:
      - 节点管理
    get:
      description: 根据ID获取节点详情
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回节点信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个节点
      tags:
      - 节点管理
    put:
      consumes:
      - application/json
      description: 更新节点配置信息
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 更新节点
      tags:
      - 节点管理
  /api/nodes/{id}/refresh:
    post:
      description: 刷新指定节点的系统信息缓存
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 刷新节点缓存
      tags:
      - 节点管理
  /api/nodes/{id}/test:
    post:
      description: 测试与LXD节点的连接状态
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 测试成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 连接失败
          schema:
            additionalProperties: true
            type: object
      summary: 测试节点连接
      tags:
      - 节点管理
...
  /nodes/{id}:
    get:
      description: 显示单个节点的详细信息和管理界面
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML页面
          schema:
            type: string
      summary: 节点详情页面
      tags:
      - 节点管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"